MINOR: quic: Move code to wakeup the timer task to avoid anti-amplication deadlock
This code was there because the timer task was not running on the same thread
as the one which parse the QUIC packets. Now that this is no more the case,
we can wake up this task directly.

Must be backported to 2.7.

(cherry picked from commit 75c8ad549002b01de0b0d139a26b356973af57a9)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>