BUG/MINOR: quic: Possible unexpected counter incrementation on send*() errors
Some counters could potentially be incremented even if send*() syscall returned
no error when ret >= 0 and ret != sz. This could be the case for instance if
a first call to send*() returned -1 with errno set to EINTR (or any previous syscall
which set errno to a non-null value) and if the next call to send*() returned
something positive and smaller than <sz>.

Must be backported to 2.7 and 2.6.

(cherry picked from commit 9fc10aff05462fe88bc117cda20d381dfb2ea9f7)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>