BUG/MINOR: mworker: fix uptime for master process
Uptime calculation for master process was incorrect as it used
<start_date> as its timestamp base time. Fix this by using the scheduler
time <start_time> for this.

The impact of this bug is minor as timestamp base time is only used for
"show proc" CLI output. it was highlighted by the following commit.
which caused a negative value to be displayed for the master process
uptime on "show proc" output.

  28360dc53f715c497fff822523f92769d8b1627d
  MEDIUM: clock: force internal time to wrap early after boot

This should be backported up to 2.0.

(cherry picked from commit 2776e775ecccc555a1d079af03d2747e825c2b96)
Signed-off-by: Willy Tarreau <w@1wt.eu>