BUG/MEDIUM: spoe: Don't set the default traget for the SPOE agent frontend
The commit d5983cef8 ("MINOR: listener: remove the useless ->default_target
field") revealed a bug in the SPOE. No default-target must be defined for
the SPOE agent frontend. SPOE applets are used on the frontend side and a
TCP connection is established on the backend side.

Because of this bug, since the commit above, the stream target is set to the
SPOE applet instead of the backend connection, leading to a spinning loop on
the applet when it is released because are unable to close the backend side.

This patch should fix the issue #2040. It only affects the 2.8-DEV but to
avoid any future bug, it should be backported to all stable versions.

(cherry picked from commit 25e36bfc228e8398171551f8c12fec29dcad0eae)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 5c8f140e5aff5f496d1694a3216ea354223cb9ef)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 6b87f11bbc07514b671317e2ab7567abddbbacd5)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>