BUG/MINOR: quic: Wrong initialization for io_cb_wakeup boolean
This bug arrives with this commit:
    982896961 MINOR: quic: split and rename qc_lstnr_pkt_rcv()
The first block of code consists in possibly setting this variable to true.
But it was already initialized to true before entering this code section.
Should be initialized to false.

Also take the opportunity to remove an unused "err" label.

Must be backported to 2.6 and 2.7.

(cherry picked from commit 8f7d22406cf35745708171dda8fa712d78a029ca)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>