BUG/MINOR: quic: Do not drop too small datagrams with Initial packets
When receiving an Initial packet a peer must drop it if the datagram is smaller
than 1200. Before this patch, this is the entire datagram which was dropped.

In such a case, drop the packet after having parsed its length.

Must be backported to 2.6 and 2.7

(cherry picked from commit 35218c6357b441142b2af19e31c8991a28b97075)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>