BUG/MINOR: quic: also send RESET_STREAM if MUX released
When the MUX is freed, the quic-conn layer may stay active until all
streams acknowledgment are processed. In this interval, if a new stream
is opened by the client, the quic-conn is thus now responsible to handle
it. This is done by the emission of a STOP_SENDING.

This process has been completed to also emit a RESET_STREAM with the
same error code H3_REQUEST_REJECTED. This is done to conform with the H3
specification to invite the client to retry its request on a new
connection.

This should be backported up to 2.6.

(cherry picked from commit 75463017123ad32ff1afe4fe95dd9088b7e9cba4)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit eb284ce40ca3425694ac5e4e1aae7b9df652f883)
[ad: replace non-existant wrapper for quic_frame alloc/dealloc by
 plain pool_alloc/pool_free]
Signed-off-by: Amaury Denoyelle <adenoyelle@haproxy.com>