BUG/MINOR: quic: acknowledge STREAM frame even if MUX is released
When the MUX is freed, the quic-conn layer may stay active until all
streams acknowledgment are processed. In this interval, if a new stream
is opened by the client, the quic-conn is thus now responsible to handle
it. This is done by the emission of a STOP_SENDING + RESET_STREAM.

Prior to this patch, the received packet was not acknowledged. This is
undesirable if the quic-conn is able to properly reject the request as
this can lead to unneeded retransmission from the client.

This must be backported up to 2.6.

(cherry picked from commit 156a89aef8c63910502b266251dc34f648a99fae)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 49c35005d7f488524b009aa8ade7e8946660bd0b)
Signed-off-by: Amaury Denoyelle <adenoyelle@haproxy.com>