MINOR: mux-h2/traces: do not log h2s pointer for dummy streams
Functions which are called with dummy streams pass it down the traces
and that leads to somewhat confusing "h2s=0x1234568(0,IDL)" for example
while the nature of the called function makes this stream useless at that
place. Better not report a random pointer, especially since it always
requires to look at the code before remembering how this should be
interpreted.

Now what we're doing is that the idle stream only prints "h2s=IDL" which
is shorter and doesn't report a pointer, closed stream do not report
anything since the stream ID 0 already implies it, and other ones are
reported normally.

This could be backported to 2.7 and 2.6 as it improves traces legibility.

(cherry picked from commit f9f4499429e678ea648b5fe587662d44402c4ea6)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>