MINOR: mux-h2/traces: add a missing TRACE_LEAVE() in h2s_frt_handle_headers()
Traces from this function would miss a TRACE_LEAVE() on the success path,
which had for consequences, 1) that it was difficult to figure where the
function was left, and 2) that we never had the allocated stream ID
clearly visible (actually the one returned by h2c_frt_stream_new() is
the right one but it's not obvious).

This can be backported to 2.7 and 2.6.

(cherry picked from commit 0d6e5d271f76a75b54b13459a3f5c86117075142)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>