BUG/MEDIUM: mworker: prevent inconsistent reload when upgrading from old versions
Previous versions ( < 1.9 ) of the master-worker process didn't had the
"HAPROXY_PROCESSES" environment variable which contains the list of
processes, fd etc.

The part which describes the master is created at first startup so if
you started the master with an old version you would never have
it.

Since patch 68836740 ("MINOR: mworker: implement a reload failure
counter"), the failedreloads member of the proc_self structure for the
master is set to 0. However if this structure does not exist, it will
result in a NULL dereference and crash the master.

This patch fixes the issue by creating the proc_self structure for the
master when it does not exist. It also shows a warning which states to
restart the master if that is the case, because we can't guarantee that
it will be working correctly.

This MUST be backported as far as 2.5, and could be backported in every
other stable branches.

(cherry picked from commit e16d32050e0c91466c2466b93c177f38f66698e2)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>
(cherry picked from commit 8648bf0d41b129cad2c42bdb807af6e8beef3a27)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>