MINOR: startup: HAPROXY_STARTUP_VERSION contains the version used to start
HAPROXY_STARTUP_VERSION: contains the version used to start, in
master-worker mode this is the version which was used to start the
master, even after updating the binary and reloading.

This patch could be backported in every version since it is useful when
debugging.

(cherry picked from commit d4c0be6b20f5365dc00221b51face2a32cf4053b)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>