BUG/MINOR: cache: Check cache entry is complete in case of Vary
Before looking for a secondary cache entry for a given request we
checked that the first entry was complete, which might prevent us from
using a valid entry if the first one with the same primary key is not
full yet.
Likewise, if the primary entry is complete but not the secondary entry
we try to use, we might end up using a partial entry from the cache as
a response.

This bug was raised in GitHub #2048.
It can be backported up to branch 2.4.

(cherry picked from commit 25917cdb12412378a80e755ffc18b5cb67c36fd2)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>