BUG/MINOR: ring: do not realign ring contents on resize
If a ring is resized, we must not zero its head since the contents
are preserved in-situ. Till now it used to work because we only resize
during boot and we emit very few data (if at all) during boot. But this
can change in the future.

This can be backported to 2.2 though no older version should notice a
difference.

(cherry picked from commit d0d85d2e364b1a8dffd605de18469c9f300aae32)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>